requirements.txt
# TAEHV训练环境依赖包
# 使用方法: pip install -r requirements.txt

# 核心深度学习框架
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# 分布式训练和优化
accelerate>=0.20.0
deepspeed>=0.9.0

# 视频生成和处理
diffusers>=0.21.0
transformers>=4.30.0
opencv-python>=4.8.0
pillow>=9.5.0

# 数据处理和可视化
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0

# 配置管理
easydict>=1.9

# 图像质量评估
lpips>=0.1.4
torchmetrics>=0.11.0
scikit-image>=0.19.0

# 系统和实用工具
psutil>=5.9.0
gpustat>=1.1.0

# 科学计算
scipy>=1.9.0

# Jupyter支持 (可选)
jupyter
ipykernel
notebook
